refactor to fix broken tests
