+++ public/app/features/dashboard-scene/pages/DashboardScenePageStateManager.ts
@@ -1,8 +1,6 @@
 import { locationUtil } from '@grafana/data';
 import { config, getBackendSrv, isFetchError, locationService } from '@grafana/runtime';
-import { updateNavIndex } from 'app/core/actions';
 import { StateManagerBase } from 'app/core/services/StateManagerBase';
-import { backendSrv } from 'app/core/services/backend_srv';
 import { default as localStorageStore } from 'app/core/store';
 import { dashboardLoaderSrv } from 'app/features/dashboard/services/DashboardLoaderSrv';
 import { getDashboardSrv } from 'app/features/dashboard/services/DashboardSrv';
@@ -10,14 +8,14 @@ import {
   DASHBOARD_FROM_LS_KEY,
   removeDashboardToFetchFromLocalStorage,
 } from 'app/features/dashboard/state/initDashboard';
-import { buildNavModel } from 'app/features/folders/state/navModel';
-import { store } from 'app/store/store';
 import { DashboardDTO, DashboardRoutes } from 'app/types';
 
 import { PanelEditor } from '../panel-edit/PanelEditor';
 import { DashboardScene } from '../scene/DashboardScene';
 import { buildNewDashboardSaveModel } from '../serialization/buildNewDashboardSaveModel';
 import { transformSaveModelToScene } from '../serialization/transformSaveModelToScene';
+
+import { updateNavModel } from './utils';
 
 export interface DashboardScenePageState {
   dashboard?: DashboardScene;
@@ -131,7 +129,7 @@ export class DashboardScenePageStateManager extends StateManagerBase<DashboardSc
 
       // Populate nav model in global store according to the folder
       if (rsp.meta.folderUid) {
-        await this.updateNavModel(rsp.meta.folderUid);
+        await updateNavModel(rsp.meta.folderUid);
       }
 
       // Do not cache new dashboards
@@ -228,15 +226,6 @@ export class DashboardScenePageStateManager extends StateManagerBase<DashboardSc
     }
 
     return null;
-  }
-
-  public async updateNavModel(folderUid: string) {
-    try {
-      const folder = await backendSrv.getFolderByUid(folderUid);
-      store.dispatch(updateNavIndex(buildNavModel(folder)));
-    } catch (err) {
-      console.warn('Error fetching parent folder', folderUid, 'for dashboard', err);
-    }
   }
 
   public clearState() {
